refactor(session): isolate snapshot validation helpers
